==> Data_preparation/1-Clear_ids.ipynb <==
import pandas as pd
import numpy as np
import os.path
from datetime import datetime
base_path = '../dataset'

original_data = os.path.join(base_path, 'original_data')
processed_data = os.path.join(base_path, 'processed_data')

candidate_items=pd.read_csv(os.path.join(original_data, "candidate_items.csv"),usecols=["item_id"]).drop_duplicates().sort_values(by="item_id").to_numpy()
candidate_items[:5]

train_sessions_items=pd.read_csv(os.path.join(original_data,"train_sessions.csv"),usecols=["item_id"]).drop_duplicates().sort_values(by="item_id").to_numpy()
train_sessions_items[:5]

train_purchases_items=pd.read_csv(os.path.join(original_data,"train_purchases.csv"),usecols=["item_id"]).drop_duplicates().sort_values(by="item_id").to_numpy()
train_purchases_items[:5]

test_sessions_items=pd.read_csv(os.path.join(original_data,"test_final_sessions.csv"),usecols=["item_id"]).drop_duplicates().sort_values(by="item_id").to_numpy()
test_sessions_items[:5]

test_sessions_leaderboard_items=pd.read_csv(os.path.join(original_data,"test_leaderboard_sessions.csv"),usecols=["item_id"]).drop_duplicates().sort_values(by="item_id").to_numpy()
test_sessions_leaderboard_items[:5]

total_test_items = np.concatenate((test_sessions_leaderboard_items, test_sessions_items))

InTestButNotInTrain=np.setdiff1d(total_test_items,train_sessions_items)
InTestButNotInTrain

print(len(InTestButNotInTrain)," items are in test sessions but not in train")

np.intersect1d(InTestButNotInTrain,train_purchases_items)
#None of those 67 items is seen in the train_purchases, so they are unknown items, should be treated in a special way

CandidateNeverBoughtInTrainingSet=np.setdiff1d(candidate_items,train_purchases_items)
CandidateNeverBoughtInTrainingSet

len(CandidateNeverBoughtInTrainingSet)

SeenInTrainSessions=np.intersect1d(CandidateNeverBoughtInTrainingSet,train_sessions_items)
SeenInTrainSessions

len(SeenInTrainSessions)

data=np.concatenate((candidate_items,train_purchases_items,train_sessions_items,test_sessions_items,test_sessions_leaderboard_items))


df=pd.DataFrame(columns=["item_id"],data=data,index=[i for i,_ in enumerate(data)]).drop_duplicates().reset_index()

df.info()

df["code"]=df.index
df["code"]+=1

del df["index"]

df.to_csv(os.path.join(processed_data,"map.csv"),index=False)

df[:19021].to_csv(os.path.join(processed_data,"map_purchases.csv"),index=False) # Mapping of only bought items

train_sessions=pd.read_csv(os.path.join(original_data,"train_sessions.csv"),parse_dates=['date'], 
                         infer_datetime_format=True,header=0)
                         
train_sessions_ids=train_sessions.sort_values("date").drop_duplicates("session_id",keep='first')["session_id"]
train_sessions_ids=train_sessions_ids.reset_index()
train_sessions_ids["index"]=train_sessions_ids.index
train_sessions_ids.to_csv(os.path.join(processed_data,"sessions_map.csv"),index=False)

test_final_sessions=pd.read_csv(os.path.join(original_data,"test_final_sessions.csv"))
test_leaderboard_sessions=pd.read_csv(os.path.join(original_data,"test_leaderboard_sessions.csv"))
test_sessions=pd.concat([test_final_sessions,test_leaderboard_sessions],axis=0)

test_sessions_ids=test_sessions.sort_values("date").drop_duplicates("session_id",keep='first')["session_id"]
test_sessions_ids=test_sessions_ids.reset_index()
test_sessions_ids["index"]=test_sessions_ids.index
test_sessions_ids.to_csv(os.path.join(processed_data,"test_sessions_map.csv"),index=False)

candidate_items=pd.read_csv(os.path.join(original_data,"candidate_items.csv"))
candidate_items=candidate_items.merge(df,on="item_id",how="left")
candidate_items["item_id"]=candidate_items["code"]
del candidate_items["code"]
candidate_items.to_csv(os.path.join(processed_data,"candidate_items_mapped.csv"),index=False)

test_final_sessions=pd.read_csv(os.path.join(original_data,"test_final_sessions.csv"))
test_final_sessions=test_final_sessions.merge(df,on="item_id",how="left")
test_final_sessions["item_id"]=test_final_sessions["code"]
del test_final_sessions["code"]
test_final_sessions=test_final_sessions.merge(test_sessions_ids,on="session_id",how="left")
test_final_sessions["session_id"]=test_final_sessions["index"]
del test_final_sessions["index"]
test_final_sessions.to_csv(os.path.join(processed_data,"test_final_sessions_full_mapped.csv"),index=False)

test_final_sessions=pd.read_csv(os.path.join(original_data,"test_final_sessions.csv"))
test_final_sessions=test_final_sessions.merge(df,on="item_id",how="left")
test_final_sessions["item_id"]=test_final_sessions["code"]
del test_final_sessions["code"]
test_final_sessions.to_csv(os.path.join(processed_data,"test_final_sessions_mapped.csv"),index=False)

test_leaderboard_sessions=pd.read_csv(os.path.join(original_data,"test_leaderboard_sessions.csv"))
test_leaderboard_sessions=test_leaderboard_sessions.merge(df,on="item_id",how="left")
test_leaderboard_sessions["item_id"]=test_leaderboard_sessions["code"]
del test_leaderboard_sessions["code"]
test_leaderboard_sessions=test_leaderboard_sessions.merge(test_sessions_ids,on="session_id",how="left")
test_leaderboard_sessions["session_id"]=test_leaderboard_sessions["index"]
del test_leaderboard_sessions["index"]
test_leaderboard_sessions.to_csv(os.path.join(processed_data,"test_leaderboard_sessions_full_mapped.csv"),index=False)

test_leaderboard_sessions=pd.read_csv(os.path.join(original_data,"test_leaderboard_sessions.csv"))
test_leaderboard_sessions=test_leaderboard_sessions.merge(df,on="item_id",how="left")
test_leaderboard_sessions["item_id"]=test_leaderboard_sessions["code"]
del test_leaderboard_sessions["code"]
test_leaderboard_sessions.to_csv(os.path.join(processed_data,"test_leaderboard_sessions_mapped.csv"),index=False)

train_purchases=pd.read_csv(os.path.join(original_data,"train_purchases.csv"))
train_purchases=train_purchases.merge(df,on="item_id",how="left")
train_purchases["item_id"]=train_purchases["code"]
del train_purchases["code"]
train_purchases=train_purchases.merge(train_sessions_ids,on="session_id",how="left")
train_purchases["session_id"]=train_purchases["index"]
del train_purchases["index"]
train_purchases.to_csv(os.path.join(processed_data,"train_purchases_mapped.csv"),index=False)

train_sessions=pd.read_csv(os.path.join(original_data,"train_sessions.csv"))
train_sessions=train_sessions.merge(df,on="item_id",how="left")
train_sessions["item_id"]=train_sessions["code"]
del train_sessions["code"]
train_sessions=train_sessions.merge(train_sessions_ids,on="session_id",how="left")
train_sessions["session_id"]=train_sessions["index"]
del train_sessions["index"]
train_sessions.to_csv(os.path.join(processed_data,"train_sessions_mapped.csv"),index=False)

item_features=pd.read_csv(os.path.join(original_data,"item_features.csv"))
item_features=item_features.merge(df,on="item_id",how="left")
item_features["item_id"]=item_features["code"]
del item_features["code"]
item_features.to_csv(os.path.join(processed_data,"item_features_mapped.csv"),index=False)

train_sessions=pd.read_csv(os.path.join(original_data,"train_sessions.csv"),parse_dates=['date'], 
                         infer_datetime_format=True,header=0)


last_month_start = datetime.strptime("01/05/21 00:00:00", '%d/%m/%y %H:%M:%S')

train_sessions_train_split=train_sessions[train_sessions["date"]<last_month_start]


train_sessions_train_split=train_sessions_train_split.merge(df,on="item_id",how="left")
train_sessions_train_split["item_id"]=train_sessions_train_split["code"]
del train_sessions_train_split["code"]
train_sessions_train_split=train_sessions_train_split.merge(train_sessions_ids,on="session_id",how="left")
train_sessions_train_split["session_id"]=train_sessions_train_split["index"]
del train_sessions_train_split["index"]
train_sessions_train_split.to_csv(os.path.join(processed_data,"train_sessions_train_split_mapped.csv"),index=False)

train_sessions_valid_split=train_sessions[train_sessions["date"]>=last_month_start]


train_sessions_valid_split=train_sessions_valid_split.merge(df,on="item_id",how="left")
train_sessions_valid_split["item_id"]=train_sessions_valid_split["code"]
del train_sessions_valid_split["code"]
train_sessions_valid_split=train_sessions_valid_split.merge(train_sessions_ids,on="session_id",how="left")
train_sessions_valid_split["session_id"]=train_sessions_valid_split["index"]
del train_sessions_valid_split["index"]
train_sessions_valid_split.to_csv(os.path.join(processed_data,"train_sessions_valid_split_mapped.csv"),index=False)

train_purchases=pd.read_csv(os.path.join(original_data,"train_purchases.csv"),parse_dates=['date'], 
                         infer_datetime_format=True,header=0)


last_month_start = datetime.strptime("01/05/21 00:00:00", '%d/%m/%y %H:%M:%S')

train_purchases_train_split=train_purchases[train_purchases["date"]<last_month_start]


train_purchases_train_split=train_purchases_train_split.merge(df,on="item_id",how="left")
train_purchases_train_split["item_id"]=train_purchases_train_split["code"]
del train_purchases_train_split["code"]
train_purchases_train_split=train_purchases_train_split.merge(train_sessions_ids,on="session_id",how="left")
train_purchases_train_split["session_id"]=train_purchases_train_split["index"]
del train_purchases_train_split["index"]
train_purchases_train_split.to_csv(os.path.join(processed_data,"train_purchases_train_split_mapped.csv"),index=False)

train_purchases_valid_split=train_purchases[train_purchases["date"]>=last_month_start]


train_purchases_valid_split=train_purchases_valid_split.merge(df,on="item_id",how="left")
train_purchases_valid_split["item_id"]=train_purchases_valid_split["code"]
del train_purchases_valid_split["code"]
train_purchases_valid_split=train_purchases_valid_split.merge(train_sessions_ids,on="session_id",how="left")
train_purchases_valid_split["session_id"]=train_purchases_valid_split["index"]
del train_purchases_valid_split["index"]
train_purchases_valid_split.to_csv(os.path.join(processed_data,"train_purchases_valid_split_mapped.csv"),index=False)

import scipy.sparse as sps

def get_URM(file='train_sessions_train_split_mapped.csv',files_directory="../dataset/processed_data", normalized=False,weight_decay=False):
    df_URM = pd.read_csv(filepath_or_buffer=os.path.join(files_directory, file), sep=',', header=0,
                         parse_dates=['date'], 
                         infer_datetime_format=True)
    
    
    df_URM['count']=1
    df_URM = df_URM.groupby(["session_id", "item_id"])["count"].sum().reset_index()
    session_id_list = df_URM['session_id'].values
    item_id_list = df_URM['item_id'].values
    
    if not normalized:
        rating_id_list = df_URM['count'].values
    else:
        rating_id_list = np.ones_like(session_id_list)


    URM = sps.csr_matrix((rating_id_list, (session_id_list, item_id_list)),shape=(1000000,23692))

    

    return URM

train_sessions_seen = get_URM()
train_bought = get_URM("train_purchases_train_split_mapped.csv")

valid_sessions_seen= get_URM("train_sessions_valid_split_mapped.csv")
valid_bought = get_URM("train_purchases_valid_split_mapped.csv")


sps.save_npz(os.path.join(processed_data, "URM_train_seen.npz"), train_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_train_bought.npz"), train_bought)

sps.save_npz(os.path.join(processed_data, "URM_valid_seen.npz"), valid_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_valid_bought.npz"), valid_bought)

sps.save_npz(os.path.join(processed_data, "URM_train_full.npz"), train_sessions_seen+train_bought)
sps.save_npz(os.path.join(processed_data, "URM_valid_full.npz"), valid_sessions_seen+valid_bought)

def get_URM_test(file='train_sessions_train_split_mapped.csv',files_directory="../dataset/processed_data", normalized=False,weight_decay=False):
    df_URM = pd.read_csv(filepath_or_buffer=os.path.join(files_directory, file), sep=',', header=0,
                         parse_dates=['date'], 
                         infer_datetime_format=True)
    
    
    df_URM['count']=1
    df_URM = df_URM.groupby(["session_id", "item_id"])["count"].sum().reset_index()
    session_id_list = df_URM['session_id'].values
    item_id_list = df_URM['item_id'].values
    
    if not normalized:
        rating_id_list = df_URM['count'].values
    else:
        rating_id_list = np.ones_like(session_id_list)


    URM = sps.csr_matrix((rating_id_list, (session_id_list, item_id_list)))

    

    return URM,np.unique(session_id_list)

test_leaderboard,leaderboard_sessions = get_URM_test(file="test_leaderboard_sessions_mapped.csv")

test_final,final_sessions = get_URM_test(file="test_final_sessions_mapped.csv")



sps.save_npz(os.path.join(processed_data, "URM_test_leaderboard.npz"), test_leaderboard)
sps.save_npz(os.path.join(processed_data, "URM_test_final.npz"), test_final)

np.save(os.path.join(processed_data, "leaderboard_sessions_ids"),leaderboard_sessions)
np.save(os.path.join(processed_data, "final_sessions_ids"),final_sessions)

def get_URM_temporal_decay(file='train_sessions_train_split_mapped.csv',files_directory="../dataset/processed_data", normalized=False,weight_decay=False,temperature=30,last_date=datetime.strptime("01/05/21 00:00:00", '%d/%m/%y %H:%M:%S')):
    df_URM = pd.read_csv(filepath_or_buffer=os.path.join(files_directory, file), sep=',', header=0,
                         parse_dates=['date'], 
                         infer_datetime_format=True)
    df_URM["diff"] = last_date-df_URM["date"]
    df_URM["days"] = df_URM["diff"].dt.days
    df_URM["days"] = df_URM["days"].apply(lambda x: max(0,x))
    df_URM["count"] = df_URM["days"].apply(lambda x: np.exp(-x/temperature))
    print(df_URM.head(10))


    df_URM = df_URM.groupby(["session_id", "item_id"])["count"].sum().reset_index()
    session_id_list = df_URM['session_id'].values
    item_id_list = df_URM['item_id'].values
    
    if not normalized:
        rating_id_list = df_URM['count'].values
    else:
        rating_id_list = np.ones_like(session_id_list)


    URM = sps.csr_matrix((rating_id_list, (session_id_list, item_id_list)),shape=(1000000,23692))

    
    print("computed")
    return URM

train_sessions_seen = get_URM_temporal_decay()
train_bought = get_URM_temporal_decay("train_purchases_train_split_mapped.csv")

valid_sessions_seen= get_URM_temporal_decay("train_sessions_valid_split_mapped.csv")
valid_bought = get_URM_temporal_decay("train_purchases_valid_split_mapped.csv")

sps.save_npz(os.path.join(processed_data, "URM_WT_train_seen.npz"), train_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WT_train_bought.npz"), train_bought)

sps.save_npz(os.path.join(processed_data, "URM_WT_valid_seen.npz"), valid_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WT_valid_bought.npz"), valid_bought)

sps.save_npz(os.path.join(processed_data, "URM_WT_train_full.npz"), train_sessions_seen+train_bought)
sps.save_npz(os.path.join(processed_data, "URM_WT_valid_full.npz"), valid_sessions_seen+valid_bought)

train_sessions_seen = get_URM_temporal_decay(last_date=datetime.strptime("01/06/21 00:00:00", '%d/%m/%y %H:%M:%S'))
train_bought = get_URM_temporal_decay("train_purchases_train_split_mapped.csv",last_date=datetime.strptime("01/06/21 00:00:00", '%d/%m/%y %H:%M:%S'))

valid_sessions_seen= get_URM_temporal_decay("train_sessions_valid_split_mapped.csv",last_date=datetime.strptime("01/06/21 00:00:00", '%d/%m/%y %H:%M:%S'))
valid_bought = get_URM_temporal_decay("train_purchases_valid_split_mapped.csv",last_date=datetime.strptime("01/06/21 00:00:00", '%d/%m/%y %H:%M:%S'))

sps.save_npz(os.path.join(processed_data, "URM_WV_train_seen.npz"), train_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WV_train_bought.npz"), train_bought)

sps.save_npz(os.path.join(processed_data, "URM_WV_valid_seen.npz"), valid_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WV_valid_bought.npz"), valid_bought)

sps.save_npz(os.path.join(processed_data, "URM_WV_train_full.npz"), train_sessions_seen+train_bought)
sps.save_npz(os.path.join(processed_data, "URM_WV_valid_full.npz"), valid_sessions_seen+valid_bought)


def get_URM_temporal_decay_within_session(file='train_sessions_train_split_mapped.csv',files_directory="../dataset/processed_data", normalized=False,weight_decay=False,temperature=5,seen=False):
    df_URM = pd.read_csv(filepath_or_buffer=os.path.join(files_directory, file), sep=',', header=0,
                         parse_dates=['date'], 
                         infer_datetime_format=True)
    df_URM["idx"] = df_URM.index

    temp=df_URM.groupby(["session_id"])["idx"].max().reset_index()
    
    temp.rename(columns={"idx":"idx_max"},inplace=True)
    df_URM=df_URM.merge(temp,on="session_id",how="left")
    df_URM["idx"]=df_URM["idx_max"]-df_URM["idx"] + 1 if seen else 0
    df_URM["count"] = df_URM["idx"].apply(lambda x: np.exp(-x/temperature))
    print(df_URM.head(10))

    df_URM = df_URM.groupby(["session_id", "item_id"])["count"].sum().reset_index()
    session_id_list = df_URM['session_id'].values
    item_id_list = df_URM['item_id'].values
    
    if not normalized:
        rating_id_list = df_URM['count'].values
    else:
        rating_id_list = np.ones_like(session_id_list)


    URM = sps.csr_matrix((rating_id_list, (session_id_list, item_id_list)),shape=(1000000,23692))

    print("computed")

    return URM

train_sessions_seen = get_URM_temporal_decay_within_session()
train_bought = get_URM_temporal_decay_within_session("train_purchases_train_split_mapped.csv")

valid_sessions_seen= get_URM_temporal_decay_within_session("train_sessions_valid_split_mapped.csv")
valid_bought = get_URM_temporal_decay_within_session("train_purchases_valid_split_mapped.csv")

sps.save_npz(os.path.join(processed_data, "URM_WW_train_seen.npz"), train_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WW_train_bought.npz"), train_bought)

sps.save_npz(os.path.join(processed_data, "URM_WW_valid_seen.npz"), valid_sessions_seen)
sps.save_npz(os.path.join(processed_data, "URM_WW_valid_bought.npz"), valid_bought)

sps.save_npz(os.path.join(processed_data, "URM_WW_train_full.npz"), train_sessions_seen+train_bought)
sps.save_npz(os.path.join(processed_data, "URM_WW_valid_full.npz"), valid_sessions_seen+valid_bought)

train_sessions_seen[:,1].sum()

def get_URM_test_temporal_decay_within_session(file='train_sessions_train_split_mapped.csv',files_directory="../dataset/processed_data", normalized=False,weight_decay=False,temperature=5):
    df_URM = pd.read_csv(filepath_or_buffer=os.path.join(files_directory, file), sep=',', header=0,
                         parse_dates=['date'], 
                         infer_datetime_format=True)
    
    
    df_URM["idx"] = df_URM.index

    temp=df_URM.groupby(["session_id"])["idx"].max().reset_index()
    
    temp.rename(columns={"idx":"idx_max"},inplace=True)
    df_URM=df_URM.merge(temp,on="session_id",how="left")
    df_URM["idx"]=df_URM["idx_max"]-df_URM["idx"] + 1 
    df_URM["count"] = df_URM["idx"].apply(lambda x: np.exp(-x/temperature))

    print(df_URM.head(10))

    df_URM = df_URM.groupby(["session_id", "item_id"])["count"].sum().reset_index()
    session_id_list = df_URM['session_id'].values
    item_id_list = df_URM['item_id'].values
    
    if not normalized:
        rating_id_list = df_URM['count'].values
    else:
        rating_id_list = np.ones_like(session_id_list)


    URM = sps.csr_matrix((rating_id_list, (session_id_list, item_id_list)))

    

    return URM

test_leaderboard = get_URM_test_temporal_decay_within_session(file="test_leaderboard_sessions_mapped.csv")

test_final = get_URM_test_temporal_decay_within_session(file="test_final_sessions_mapped.csv")

sps.save_npz(os.path.join(processed_data, "URM_WW_test_leaderboard.npz"), test_leaderboard)
sps.save_npz(os.path.join(processed_data, "URM_WW_test_final.npz"), test_final)

test_leaderboard[:,1].sum()

import tensorflow as tf

assert tf.__version__=="2.8.2","tf_version is wrong"
